refactor: [MON-177] pages 폴더구조 정리
