--- src/pages/EditArticlePage/index.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const EditArticlePage = () => <div />;
-
-export default EditArticlePage;
