--- src/pages/MyChannelPage/index.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const MyChannelPage = () => <div />;
-
-export default MyChannelPage;
