+++ src/pages/article/WriteArticlePage.jsx
@@ -8,8 +8,8 @@ import {
 import { useForm } from '@hooks';
 import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
-import { postArticle } from '../../apis/article';
-import jsonBlob from '../../utils/createJsonBlob';
+import { postArticle } from '@apis/article';
+import jsonBlob from '@utils/createJsonBlob';
 
 const WriteArticlePage = ({ match, history }) => {
   const { id } = match.params;
