+++ src/pages/article/index.jsx
@@ -1,0 +1,3 @@
+export { default as ArticleDetailPage } from './ArticleDetailPage';
+export { default as EditArticlePage } from './EditArticlePage';
+export { default as WriteArticlePage } from './WriteArticlePage';
