+++ src/pages/auth/index.jsx
@@ -1,0 +1,2 @@
+export { default as SignUpPage } from './SignUpPage';
+export { default as SignInPage } from './SignInPage';
