+++ src/pages/channel/index.jsx
@@ -1,0 +1,1 @@
+export { default as ChannelPage } from './ChannelPage';
