+++ src/pages/general/index.jsx
@@ -1,0 +1,3 @@
+export { default as HomePage } from './HomePage';
+export { default as NotFoundPage } from './NotFoundPage';
+export { default as SearchPage } from './SearchPage';
