+++ src/pages/purchase/index.jsx
@@ -1,0 +1,2 @@
+export { default as PurchaseHistoryPage } from './PurchaseHistoryPage';
+export { default as PurchasePage } from './PurchasePage';
