+++ src/pages/series/EditSeriesPage.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import styled from '@emotion/styled';
 import PropTypes from 'prop-types';
 import {
@@ -13,14 +13,15 @@ import {
 } from '@components';
 import { useForm } from '@hooks';
 import calculateLaterDate from '@utils/calculateLaterDate ';
-import jsonBlob from '@utils/createJsonBlob';
 import convertSeriesInputName from '@utils/convertSeriesInputName';
-import { postSeries } from '../../apis/series';
+import { putSeries, getSeriesDetail, putSeriesImage } from '@apis/series';
+import jsonBlob from '@utils/createJsonBlob';
 
-const WriteSeriesPage = ({ history }) => {
-  const [file, setFile] = useState();
+const EditSeriesPage = ({ match, history }) => {
+  const { id } = match.params;
+  const [file, setFile] = useState(null);
   const [checkedInputs, setCheckedInputs] = useState([]);
-  const { values, handleChange, handleSubmit } = useForm({
+  const { values, setValues, handleChange, handleSubmit } = useForm({
     initialValues: {
       title: '',
       introduceText: '',
@@ -33,36 +34,39 @@ const WriteSeriesPage = ({ history }) => {
       category: '',
       uploadTime: '',
       articleCount: '',
+      thumbnail: '',
     },
 
     onSubmit: async values => {
-      if (!file) {
-        alert('이미지를 업로드 해주세요!');
-        return;
-      }
-
-      if (checkedInputs.length === 0) {
-        alert('요일을 선택해주세요!');
-        return;
-      }
-
       try {
-        const requestData = {
-          ...values,
+        const request = {
+          writeId: values.writeId,
+          title: values.title,
+          introduceText: values.introduceText,
+          introduceSentence: values.introduceSentence,
           uploadDate: checkedInputs,
-          articleCount: Number(values.articleCount),
-          price: Number(values.price),
+          uploadTime: values.uploadTime,
         };
 
-        const formData = new FormData();
-        formData.append('file', file);
-        formData.append('request', jsonBlob(requestData));
-
-        const response = await postSeries({
-          data: formData,
+        const textResponse = await putSeries({
+          data: jsonBlob(request),
+          params: id,
         });
-        const { seriesId } = response.data;
-        seriesId && history.push(`/series/${seriesId}`);
+
+        if (file) {
+          const fileForm = new FormData();
+          fileForm.append('file', file);
+
+          const fileResponse = await putSeriesImage({
+            data: fileForm,
+            params: id,
+          });
+          textResponse.status === 200 &&
+            fileResponse.status === 200 &&
+            history.push(`/series/${id}`);
+        } else {
+          textResponse.status === 200 && history.push(`/series/${id}`);
+        }
       } catch (error) {
         alert(error);
       }
@@ -71,14 +75,47 @@ const WriteSeriesPage = ({ history }) => {
       const newErrors = {};
       for (const key in values) {
         if (!values[key]) {
-          newErrors.empty = `${convertSeriesInputName(key)}을 입력해주세요!`;
-          alert(`${convertSeriesInputName(key)}을 입력해주세요!`);
+          newErrors.empty = `${convertSeriesInputName(key)}를 입력해주세요!`;
+          alert(`${convertSeriesInputName(key)}를 입력해주세요!`);
           break;
         }
+      }
+      if (values.uploadDate.length !== checkedInputs.length) {
+        newErrors.dayLength = '요일 수가 일치하지 않습니다!';
+        alert('요일 수가 일치하지 않습니다!');
       }
       return newErrors;
     },
   });
+
+  const init = async id => {
+    const response = await getSeriesDetail({
+      params: id,
+    });
+    const { series, upload, subscribe, category, writer } = response.data;
+
+    setValues({
+      writeId: writer.id,
+      title: series.title,
+      introduceText: series.introduceText,
+      introduceSentence: series.introduceSentence,
+      price: series.price,
+      subscribeStartDate: subscribe.startDate,
+      subscribeEndDate: subscribe.endDate,
+      seriesStartDate: series.startDate,
+      seriesEndDate: series.endDate,
+      category,
+      uploadTime: upload.time,
+      articleCount: series.articleCount,
+      uploadDate: upload.date,
+      thumbnail: series.thumbnail,
+    });
+    setCheckedInputs(upload.date);
+  };
+
+  useEffect(() => {
+    id && init(id);
+  }, []);
 
   const handleChangefile = file => {
     file && setFile(file);
@@ -91,15 +128,15 @@ const WriteSeriesPage = ({ history }) => {
       setCheckedInputs(checkedInputs.filter(el => el !== value));
     }
   };
-
   return (
-    <StyledWrapper>
+    <StyledWrapper styled={{ padding: '2rem 0' }}>
       <form onSubmit={handleSubmit}>
         <Section>
           <Radio
             names={['poem', 'novel', 'interview', 'essay', 'critique', 'etc']}
             onChange={handleChange}
             checkedButton={values.category}
+            disabled={!!id}
             title="카테고리"
           />
         </Section>
@@ -113,7 +150,11 @@ const WriteSeriesPage = ({ history }) => {
         </Section>
 
         <Section>
-          <ImageUpload onChange={handleChangefile} title="이미지 업로드" />
+          <ImageUpload
+            onChange={handleChangefile}
+            title="이미지 업로드"
+            url={values.thumbnail}
+          />
         </Section>
 
         <Section>
@@ -124,6 +165,7 @@ const WriteSeriesPage = ({ history }) => {
             name="price"
             onChange={handleChange}
             min={0}
+            disabled={!!id}
           />
         </Section>
 
@@ -137,6 +179,7 @@ const WriteSeriesPage = ({ history }) => {
             endValue={values.subscribeEndDate}
             endMin={calculateLaterDate(values.subscribeStartDate, 1)}
             onChange={handleChange}
+            pageParam={id}
           />
         </Section>
 
@@ -150,6 +193,7 @@ const WriteSeriesPage = ({ history }) => {
             endValue={values.seriesEndDate}
             endMin={calculateLaterDate(values.seriesStartDate, 1)}
             onChange={handleChange}
+            pageParam={id}
           />
         </Section>
 
@@ -171,6 +215,7 @@ const WriteSeriesPage = ({ history }) => {
             value={values.articleCount}
             onChange={handleChange}
             min={1}
+            disabled={!!id}
           />
         </Section>
 
@@ -197,11 +242,12 @@ const WriteSeriesPage = ({ history }) => {
   );
 };
 
-WriteSeriesPage.propTypes = {
+EditSeriesPage.propTypes = {
+  match: PropTypes.object.isRequired,
   history: PropTypes.object.isRequired,
 };
 
-export default WriteSeriesPage;
+export default EditSeriesPage;
 
 const StyledWrapper = styled(Wrapper)`
   padding: 9rem 0 4rem 0;
