+++ src/pages/series/SeriesDetailPage.jsx
