+++ src/pages/series/index.jsx
@@ -1,0 +1,4 @@
+export { default as EditSeriesPage } from './EditSeriesPage';
+export { default as SeriesListPage } from './SeriesListPage';
+export { default as WriteSeriesPage } from './WriteSeriesPage';
+export { default as SeriesDetailPage } from './SeriesDetailPage';
