+++ src/pages/user/index.jsx
@@ -1,0 +1,3 @@
+export { default as EditMyInfoPage } from './EditMyInfoPage';
+export { default as MyInfoPage } from './MyInfoPage';
+export { default as WriteListPage } from './WriteListPage';
